requirements.txt
flask~=3.0.0
pymongo~=4.6.0
gunicorn~=21.2.0
Pillow~=10.1.0